fixs

+++ services/comments/delete.ts
@@ -1,6 +1,7 @@
 import { PrismaClient } from "@prisma/client";
+import { JwtPayload } from "jsonwebtoken";
 import verify from "../verify";
-const prisma = new PrismaClient()
+const prisma = new PrismaClient();
 
 interface IComment {
   id: number;
@@ -28,18 +29,19 @@ interface IUser {
 }
 
 export default async function deleteComment({comment, token}: IBody) {
-  const user: IUser = await verify(token)
+  const user: IUser | boolean | string | JwtPayload = await verify(token);
 
-  if(!user) return "mistake"
-  if(user.status === "BANNED") return "mistake"
+  if(typeof user === "boolean" || typeof user === "string") return "mistake";
+  if(!user) return "mistake";
+  if(user.status === "BANNED") return "mistake";
 
-  if(!comment) return "mistake"
+  if(!comment) return "mistake";
 
-  if(!user.role.includes("ADMIN")) return "mistake"
+  if(!user.role.includes("ADMIN")) return "mistake";
   await prisma.comment.delete({
     where: {
       id: comment.id
     }
-  })
-  return "succesful"
+  });
+  return "succesful";
 }
